Refactor commands into separate class
